requirements.txt
asgiref==3.4.1
Django==2.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
